# src/main/resources/swagger.json
{"swagger":"2.0","info":{"description":"Employee Management REST API","version":"1.0.0","title":"Spring Boot REST API","contact":{"name":"Ramesh Fadatare","url":"www.javaguides.net","email":"ramesh24fadatare@gmail.com"},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0.html"}},"host":"localhost:7040","basePath":"/","tags":[{"name":"utilisateur-controller","description":"Ressources permettant de gérer des utilisateurs (CRUD)"}],"paths":{"/create":{"post":{"tags":["utilisateur-controller"],"summary":"Créé un utilisateur","operationId":"createUtilisateurUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"query","required":false,"type":"integer","format":"int32"},{"name":"nom","in":"query","required":false,"type":"string"},{"name":"prenom","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Utilisateur"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/delete/{id}":{"delete":{"tags":["utilisateur-controller"],"summary":"Supprime un utilisateur","operationId":"deleteUtilisateurByIdUsingDELETE","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Identifiant d'un utilisateur (ex 1,2,3...)","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Utilisateur"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/get":{"get":{"tags":["utilisateur-controller"],"summary":"Récupère la liste des utilisateurs","operationId":"getAllUtilisateursUsingGET","produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"object"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/get/{id}":{"get":{"tags":["utilisateur-controller"],"summary":"Récupère un utilisateur","operationId":"getUtilisateurByIdUsingGET","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Identifiant d'un utilisateur (ex 1,2,3...)","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Utilisateur"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/update/{id}":{"put":{"tags":["utilisateur-controller"],"summary":"Met à jour un utilisateur","operationId":"updateUtilisateurUsingPUT","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Identifiant d'un utilisateur (ex 1,2,3...)","required":true,"type":"integer","format":"int32"},{"name":"id","in":"query","required":false,"type":"integer","format":"int32"},{"name":"nom","in":"query","required":false,"type":"string"},{"name":"prenom","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Utilisateur"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"definitions":{"Utilisateur":{"type":"object","properties":{"id":{"type":"integer","format":"int32","description":"Identifiant de l'utilisateur","allowEmptyValue":false},"nom":{"type":"string","description":"Nom de l'utilisateur","allowEmptyValue":false},"prenom":{"type":"string","description":"Prénom de l'utilisateur","allowEmptyValue":false}},"title":"Utilisateur"}}}